Changed counts from size_t to uint32_t
git-svn-id: svn+ssh://subversion.munts.net/src/trunk/ARM@3024 06ee2f96-5407-11de-8bab-f3d7589a4122

+++ stm32/spi.h
@@ -19,6 +19,6 @@ int spimaster_init(uint32_t port,
 
 int spimaster_transfer(uint32_t port,
                        uint8_t *txbuf,
-                       size_t txcount,
+                       uint32_t txcount,
                        uint8_t *rxbuf,
-                       size_t rxcount);
+                       uint32_t rxcount);
